1. Converting all medicine articles activities and classes to fitness articles and classes
