+++ app/src/main/res/layout/activity_home.xml
@@ -124,7 +124,7 @@
             </LinearLayout>
         </androidx.cardview.widget.CardView>
         <androidx.cardview.widget.CardView
-            android:id="@+id/cardHealthDoctor"
+            android:id="@+id/cardFitnessArticle"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_row="1"
@@ -150,7 +150,7 @@
                 <TextView
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="HEALTH ARTICLES"
+                    android:text="FITNESS ARTICLES"
                     android:textAlignment="center"
                     android:textColorHint="@color/black"
                     android:textStyle="bold"/>
