+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">Healthcare</string>
+    <string name="app_name">Fit And Health</string>
     <string name="username">Username...</string>
 </resources>
